refactor: made a separated package for http server
moved the http server into its own package as tcp now also has its own package
