+++ internal/server/http/http_bench_test.go
@@ -1,4 +1,4 @@
-package server
+package http
 
 import (
 	"github.com/fakefloordiv/indigo/http"
