+++ internal/server/http/states.go
@@ -1,4 +1,4 @@
-package server
+package http
 
 type serverState uint8
